Extract FailureInjector interface
