+++ core/trino-main/src/main/java/io/trino/execution/TestingFailureInjector.java
@@ -17,29 +17,18 @@ import com.google.common.cache.CacheBuilder;
 import com.google.inject.Inject;
 import io.airlift.units.Duration;
 import io.trino.cache.NonEvictableCache;
-import io.trino.spi.ErrorCode;
-import io.trino.spi.ErrorCodeSupplier;
 import io.trino.spi.ErrorType;
-import io.trino.spi.TrinoException;
 
 import java.util.Objects;
 import java.util.Optional;
 
-import static com.google.common.base.MoreObjects.toStringHelper;
-import static com.google.common.base.Preconditions.checkArgument;
 import static io.trino.cache.SafeCaches.buildNonEvictableCache;
-import static io.trino.execution.TestingFailureInjector.InjectedFailureType.TASK_FAILURE;
-import static io.trino.spi.ErrorType.EXTERNAL;
-import static io.trino.spi.ErrorType.INSUFFICIENT_RESOURCES;
-import static io.trino.spi.ErrorType.INTERNAL_ERROR;
-import static io.trino.spi.ErrorType.USER_ERROR;
 import static java.util.Objects.requireNonNull;
 import static java.util.concurrent.TimeUnit.MILLISECONDS;
 
 public class TestingFailureInjector
+        implements FailureInjector
 {
-    public static final String FAILURE_INJECTION_MESSAGE = "This error is injected by the failure injection service";
-
     private final NonEvictableCache<Key, InjectedFailure> failures;
     private final Duration requestTimeout;
 
@@ -58,6 +47,7 @@ public class TestingFailureInjector
         this.requestTimeout = requireNonNull(requestTimeout, "requestTimeout is null");
     }
 
+    @Override
     public void injectTaskFailure(
             String traceToken,
             int stageId,
@@ -69,6 +59,7 @@ public class TestingFailureInjector
         failures.put(new Key(traceToken, stageId, partitionId, attemptId), new InjectedFailure(injectionType, errorType));
     }
 
+    @Override
     public Optional<InjectedFailure> getInjectedFailure(
             String traceToken,
             int stageId,
@@ -81,6 +72,7 @@ public class TestingFailureInjector
         return Optional.ofNullable(failures.getIfPresent(new Key(traceToken, stageId, partitionId, attemptId)));
     }
 
+    @Override
     public Duration getRequestTimeout()
     {
         return requestTimeout;
@@ -118,91 +110,6 @@ public class TestingFailureInjector
         public int hashCode()
         {
             return Objects.hash(traceToken, stageId, partitionId, attemptId);
-        }
-    }
-
-    public enum InjectedFailureType
-    {
-        TASK_MANAGEMENT_REQUEST_FAILURE,
-        TASK_MANAGEMENT_REQUEST_TIMEOUT,
-        TASK_GET_RESULTS_REQUEST_FAILURE,
-        TASK_GET_RESULTS_REQUEST_TIMEOUT,
-        TASK_FAILURE,
-    }
-
-    public static class InjectedFailure
-    {
-        private final InjectedFailureType injectedFailureType;
-        private final Optional<ErrorType> taskFailureErrorType;
-
-        public InjectedFailure(InjectedFailureType injectedFailureType, Optional<ErrorType> taskFailureErrorType)
-        {
-            this.injectedFailureType = requireNonNull(injectedFailureType, "injectedFailureType is null");
-            this.taskFailureErrorType = requireNonNull(taskFailureErrorType, "taskFailureErrorType is null");
-            if (injectedFailureType == TASK_FAILURE) {
-                checkArgument(taskFailureErrorType.isPresent(), "error type must be present when injection type is task failure");
-            }
-            else {
-                checkArgument(taskFailureErrorType.isEmpty(), "error type must not be present when injection type is not task failure");
-            }
-        }
-
-        public InjectedFailureType getInjectedFailureType()
-        {
-            return injectedFailureType;
-        }
-
-        public ErrorType getTaskFailureErrorType()
-        {
-            return taskFailureErrorType.orElseThrow(() -> new IllegalStateException("this method must only be called for failure type of TASK_FAILURE"));
-        }
-
-        public Throwable getTaskFailureException()
-        {
-            ErrorType errorType = getTaskFailureErrorType();
-            return new TrinoException(InjectedErrorCode.getErrorCode(errorType), FAILURE_INJECTION_MESSAGE);
-        }
-
-        @Override
-        public String toString()
-        {
-            return toStringHelper(this)
-                    .add("injectedFailureType", injectedFailureType)
-                    .add("taskFailureErrorType", taskFailureErrorType)
-                    .toString();
-        }
-    }
-
-    public enum InjectedErrorCode
-            implements ErrorCodeSupplier
-    {
-        INJECTED_USER_ERROR(1, USER_ERROR),
-        INJECTED_INTERNAL_ERROR(2, INTERNAL_ERROR),
-        INJECTED_INSUFFICIENT_RESOURCES_ERROR(3, INSUFFICIENT_RESOURCES),
-        INJECTED_EXTERNAL_ERROR(4, EXTERNAL),
-        /**/;
-
-        private final ErrorCode errorCode;
-
-        InjectedErrorCode(int code, ErrorType type)
-        {
-            errorCode = new ErrorCode(code + 0x30000, name(), type);
-        }
-
-        @Override
-        public ErrorCode toErrorCode()
-        {
-            return errorCode;
-        }
-
-        public static InjectedErrorCode getErrorCode(ErrorType errorType)
-        {
-            for (InjectedErrorCode code : values()) {
-                if (code.toErrorCode().getType() == errorType) {
-                    return code;
-                }
-            }
-            throw new IllegalArgumentException("unexpected error type: " + errorType);
         }
     }
 }
