+++ testing/trino-testing/src/main/java/io/trino/testing/BaseFailureRecoveryTest.java
@@ -21,7 +21,7 @@ import io.airlift.units.Duration;
 import io.trino.Session;
 import io.trino.client.StageStats;
 import io.trino.client.StatementStats;
-import io.trino.execution.TestingFailureInjector.InjectedFailureType;
+import io.trino.execution.FailureInjector.InjectedFailureType;
 import io.trino.operator.RetryPolicy;
 import io.trino.spi.ErrorType;
 import io.trino.spi.QueryId;
@@ -49,12 +49,12 @@ import static com.google.common.base.Functions.identity;
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.collect.ImmutableList.toImmutableList;
 import static com.google.common.collect.Streams.stream;
-import static io.trino.execution.TestingFailureInjector.FAILURE_INJECTION_MESSAGE;
-import static io.trino.execution.TestingFailureInjector.InjectedFailureType.TASK_FAILURE;
-import static io.trino.execution.TestingFailureInjector.InjectedFailureType.TASK_GET_RESULTS_REQUEST_FAILURE;
-import static io.trino.execution.TestingFailureInjector.InjectedFailureType.TASK_GET_RESULTS_REQUEST_TIMEOUT;
-import static io.trino.execution.TestingFailureInjector.InjectedFailureType.TASK_MANAGEMENT_REQUEST_FAILURE;
-import static io.trino.execution.TestingFailureInjector.InjectedFailureType.TASK_MANAGEMENT_REQUEST_TIMEOUT;
+import static io.trino.execution.FailureInjector.FAILURE_INJECTION_MESSAGE;
+import static io.trino.execution.FailureInjector.InjectedFailureType.TASK_FAILURE;
+import static io.trino.execution.FailureInjector.InjectedFailureType.TASK_GET_RESULTS_REQUEST_FAILURE;
+import static io.trino.execution.FailureInjector.InjectedFailureType.TASK_GET_RESULTS_REQUEST_TIMEOUT;
+import static io.trino.execution.FailureInjector.InjectedFailureType.TASK_MANAGEMENT_REQUEST_FAILURE;
+import static io.trino.execution.FailureInjector.InjectedFailureType.TASK_MANAGEMENT_REQUEST_TIMEOUT;
 import static io.trino.plugin.base.TemporaryTables.temporaryTableNamePrefix;
 import static io.trino.testing.QueryAssertions.assertEqualsIgnoreOrder;
 import static io.trino.testing.TestingNames.randomNameSuffix;
