+++ testing/trino-testing/src/main/java/io/trino/testing/DistributedQueryRunner.java
@@ -31,8 +31,8 @@ import io.trino.Session;
 import io.trino.Session.SessionBuilder;
 import io.trino.connector.CoordinatorDynamicCatalogManager;
 import io.trino.cost.StatsCalculator;
+import io.trino.execution.FailureInjector.InjectedFailureType;
 import io.trino.execution.QueryManager;
-import io.trino.execution.TestingFailureInjector.InjectedFailureType;
 import io.trino.execution.warnings.WarningCollector;
 import io.trino.metadata.AllNodes;
 import io.trino.metadata.FunctionBundle;
