+++ testing/trino-testing/src/main/java/io/trino/testing/ExtendedFailureRecoveryTest.java
@@ -32,10 +32,10 @@ import static com.google.common.collect.Iterables.getOnlyElement;
 import static io.trino.SystemSessionProperties.ENABLE_DYNAMIC_FILTERING;
 import static io.trino.SystemSessionProperties.JOIN_DISTRIBUTION_TYPE;
 import static io.trino.SystemSessionProperties.JOIN_REORDERING_STRATEGY;
-import static io.trino.execution.TestingFailureInjector.FAILURE_INJECTION_MESSAGE;
-import static io.trino.execution.TestingFailureInjector.InjectedFailureType.TASK_FAILURE;
-import static io.trino.execution.TestingFailureInjector.InjectedFailureType.TASK_MANAGEMENT_REQUEST_FAILURE;
-import static io.trino.execution.TestingFailureInjector.InjectedFailureType.TASK_MANAGEMENT_REQUEST_TIMEOUT;
+import static io.trino.execution.FailureInjector.FAILURE_INJECTION_MESSAGE;
+import static io.trino.execution.FailureInjector.InjectedFailureType.TASK_FAILURE;
+import static io.trino.execution.FailureInjector.InjectedFailureType.TASK_MANAGEMENT_REQUEST_FAILURE;
+import static io.trino.execution.FailureInjector.InjectedFailureType.TASK_MANAGEMENT_REQUEST_TIMEOUT;
 import static io.trino.spi.predicate.Domain.singleValue;
 import static io.trino.spi.type.BigintType.BIGINT;
 import static io.trino.sql.planner.OptimizerConfig.JoinDistributionType.PARTITIONED;
